feat: extract CLI command handlers into grouped modules
Move all command handler functions from cli.ts into three modules:
- commands/verify.ts (cmdVerify, cmdScore, cmdLayerTest)
- commands/analytics.ts (cmdStatus, cmdTrend, cmdDiff, cmdBaseline, cmdHistory, cmdInsights, cmdTasks, cmdReport)
- commands/utility.ts (cmdInit, cmdInject, cmdSuggest, cmdContext, cmdResultUpdate, cmdConfigCheck, cmdClean, cmdSkillLint, cmdLayers)

Each function now takes (args, repoRoot) parameters instead of using module globals.
cli.ts imports from commands and dispatches. All 61 existing tests pass unchanged.

Closes #98
