+++ .canductor/results.tsv
@@ -34,3 +34,4 @@ guard-ref	2026-03-23T23:13:27.937Z	38	block	typecheck:100,tests:0,code_quality:0
 89	2026-03-24T05:08:10.995Z	99	auto_merge	typecheck:100,tests:100,code_quality:95	merged	Verified 89
 90	2026-03-24T05:12:35.350Z	99	auto_merge	typecheck:100,tests:100,code_quality:94	merged	Verified 90
 93	2026-03-24T05:14:41.976Z	98	auto_merge	typecheck:100,tests:100,code_quality:93	merged	Verified 93
+98	2026-03-24T05:29:38.215Z	99	auto_merge	typecheck:100,tests:100,code_quality:96	pending	Verified 98
