+++ .canductor/tasks.tsv
@@ -7,3 +7,7 @@ module	.claude/templates/module.md	#90	1	none	2026-03-24T05:12:41Z
 test	.claude/templates/test.md	#90	1	none	2026-03-24T05:12:41Z
 new-cli-command	.claude/patterns/new-cli-command.md	#93	1	none	2026-03-24T05:14:47Z
 test	.claude/templates/test.md	#93	1	none	2026-03-24T05:14:47Z
+module	.claude/templates/module.md	#98	1	none	2026-03-24T05:29:51Z
+module	.claude/templates/module.md	#98	1	none	2026-03-24T05:29:51Z
+module	.claude/templates/module.md	#98	1	none	2026-03-24T05:29:51Z
+refactor	-	#98	1	none	2026-03-24T05:29:51Z
